Use macOS for the formatting check.
Observation of the Azure CI pipelines indicates that installation and
build on the macOS ('Darwin') agent is substantially (~40%) faster than
on the Ubuntu ('Linux') agent. While the formatting check is not long,
and the real blocker for us is currently Windows, we can still drop the
Ubuntu time slightly by moving it to the macOS run, which may at least
produce failures faster.