[ci docs] Use the cargo flag for --document-private-items
It appears that setting RUSTDOCFLAGS="--document-private-items" can cause
rustdoc to fail because the setting is set twice, once by cargo and again
by the environment variable. This results in the docs builds failing to
execute.

Updated to use the cargo flag, which ensures that the flag is only passed
to rustdoc once.

Also updated the build-api-docs CI task to include recently added crates.