Support creating completions in new directories
Previously, we were failing when any part of the path did not exist,
and reporting the wrong error to boot. Here, we attempt to create the
directory instead, informing the user that we are changing their file
system. Adds a new error scenario to cover a failure (e.g. a
permissions error on creation), and a success message after writing
the completion file.

Fixes #645