Update cli_platform tests to use 'const' correctly
The tests of the platform resolution for the various 'volta run'
scenarios were using a lazily-loaded 'const' value. Clippy raised
some warnings about those, as 'const' values are actually duplicated
for each location they are used, so the interior mutability wasn't
helpful.

It seems that the `Lazy` behavior was actually a way to work around
the fact that `Version` doesn't have a 'const' constructor, however
all of the fields are `pub`, so we can construct the values directly.
This allows us to have a true `const` value and use it in each test
without needing extra interior mutability or cloning.