Update Node fixtures to re-use Node filename logic
The current Node test fixtures and smoke tests use duplicated
logic to determine the file name for Node distributions. However,
that logic was never updated to reflect the fallback behavior
that we use for Apple Silicon machines. This means that when
running tests on those machines, we have issues with the names
not lining up correctly.

This became a higher priority issue lately as GitHub Actions
switched their default 'macos-latest' runners to use M1 Macs,
causing our CI to fail.