Update shim behavior on Windows to use scripts
Rather than using symlinks, which require administrator privileges
or developer mode on Windows, we can instead use small scripts that
call out to 'volta run' to actually handle the execution of 3rd
party binaries.

This also adds a migration to a new v4 layout, which allows Windows
installs to migrate immediately to using the script-based shims.