debugging: add a bunch of tracing around `archive`
This is not really intended to be merged as is, but it should allow
folks to do some testing and see if we can get more info for #1744.

In the interest of *possibly* being able to build on this for merging it
if we find that useful in the future, I used `trace!` rather than
`debug!` for this extreme level of output, and added a `--very-verbose`
flag for this mode.