Remove extra GET for tarball uncompressed size
Currently, we make two overlapping GET requests when fetching a tarball:
First, the main request to get the headers and the data stream, and then
a second request to attempt to fetch the ISIZE field which holds the
uncompressed size. Getting the uncompressed size lets us use the stream
data after uncompressing to populate our progress bar for tarballs.

However, the extra request causes issues for some users, where the
second request for the ISIZE field hangs during the TLS handshake. By
removing the uncompressed size entirely and connecting the progress bar
to the compressed read data, we can avoid the extra request entirely.