Remove early-exit check_architecture function from installer
Early on, we added a fast-fail function `check_architecture` to the installer.
This allowed us to detect an unsupported architecture and fail early, rather
than continuing and appearing to succeed only for the app to fail.

As of 2.0.0, however, we now support both x64 & ARM on all platforms, so we
check the architecture as part of the normal install process. This means that
the early exit is no longer necessary—an unsupported architecture will still
be caught by the `parse_os_info` function before any download occurs.