Update `check_shim_reachable` to correctly handle Windows
We have a helper utility `check_shim_reachable` that notifies the user if the shim
for a tool they just installed is on their PATH and whether it's shadowed. However,
that check currently assumes that all shims are in `VoltaHome::shim_dir`. Unfortunately,
on Windows that isn't the case—we install the default shims alongside Volta itself
inside of `Program Files`.

To prevent showing an unnecessary message to Windows users, we should properly check
both Volta directories on Windows when validating whether a shim is properly available
on the Path.