Move pin suggestion CTA to the project version message
Currently, the call-to-action suggestion to use `volta pin` is printed
as part of the "success" message on every call to `volta install`.
However, that message is confusing when you aren't in a project.
Additionally, it's only necessary if the current project itself has a
pinned version of the tool.

We actually already have a message sent after installing if the project
has an overlapping version: a note that the local version is different.
This message seems like the ideal place to have the CTA, so that it lets
users know what they can do to update the local version in case that's
what they intended.