Use lockfile when installing cargo-wix in CI
When installing a package using `cargo install`, by default it ignores
the lockfile for that package and installs the latest relevant
dependency versions. This can be overridden by using the `--locked`
flag.

Using an unlocked install is causing issues in CI, because we have our
Rust version pinned to a version that is older than the MSRV for some of
`cargo-wix`'s dependencies. Switching to a locked install means that the
dependency versions will be repeatable and so we shouldn't have issues
with Rust version mismatches.